1.更改了文件命名方式 2.添加了Function、Exp、Square类
