+++ LM_From_None/Mytorch/Variable.py
@@ -1,0 +1,13 @@
+import numpy as np
+
+class Variable:#定义深度学习的变量类
+    def __init__(self, data):
+        self.data = data
+
+if __name__ == '__main__':
+#通过numpy.array测试
+    data = np.array(1.0)
+    x = Variable(data)
+    print(x.data)
+    x.data = np.array(2.0)
+    print(x.data)
